test: testes da classe BonusService

+++ src/models/Funcionario.java
@@ -5,26 +5,26 @@ import java.time.LocalDate;
 
 public class Funcionario {
 
-	private String nome;
-	private LocalDate dataAdmissao;
-	private BigDecimal salario;
+    private String nome;
+    private LocalDate dataAdmissao;
+    private double salario;
 
-	public Funcionario(String nome, LocalDate dataAdmissao, BigDecimal salario) {
-		this.nome = nome;
-		this.dataAdmissao = dataAdmissao;
-		this.salario = salario;
-	}
+    public Funcionario(String nome, LocalDate dataAdmissao, double salario) {
+        this.nome = nome;
+        this.dataAdmissao = dataAdmissao;
+        this.salario = salario;
+    }
 
-	public String getNome() {
-		return nome;
-	}
+    public String getNome() {
+        return nome;
+    }
 
-	public LocalDate getDataAdmissao() {
-		return dataAdmissao;
-	}
+    public LocalDate getDataAdmissao() {
+        return dataAdmissao;
+    }
 
-	public BigDecimal getSalario() {
-		return salario;
-	}
+    public double getSalario() {
+        return salario;
+    }
 
 }
